Split player-state.spec.ts into per-domain test files
Redistribute the oversized player-state.spec.ts (~710 lines, over the 600-line
test cap) into a player-state-tests/ directory (firebase-to-player-state,
werewolf-core, werewolf-roles specs + shared werewolf-fixtures), preserving
every existing assertion. Same test count runs and passes.
